+++ py/desimodel/focalplane.py
@@ -14,6 +14,12 @@ import numpy as np
 from astropy.io import fits
 import astropy.units as u
 
+# Define this here to avoid a problem with Sphinx compilation.
+try:
+    default_offset = 10.886*u.um
+except TypeError:
+    default_offset = 10.886
+
 
 def generate_random_vector_field(rms, exponent, n, seed=None, smoothing=0.05):
     """Generate a pair dx, dy of 2D Gaussian random field.
@@ -68,7 +74,7 @@ def generate_random_vector_field(rms, exponent, n, seed=None, smoothing=0.05):
     return dx, dy
 
 
-def generate_random_centroid_offsets(rms_offset=10.886*u.um, seed=123):
+def generate_random_centroid_offsets(rms_offset=default_offset, seed=123):
     """Generate random centroid offsets.
 
     Calls :func:`generate_random_vector_field` to generate offsets with
@@ -82,9 +88,9 @@ def generate_random_centroid_offsets(rms_offset=10.886*u.um, seed=123):
 
     Parameters
     ----------
-    rms_offset : astropy quantity
+    rms_offset : :class:`astropy.Quantity` instance.
         RMS that the generated offsets should have, including units.
-    seed : int
+    seed : :class:`int`
         Random number seed to use. Generated offsets should be portable
         across python versions and platforms.
 
@@ -121,9 +127,8 @@ class FocalPlane(object):
         self._fiberpos_file = os.path.join(os.environ['DESIMODEL'],
                                            'data', 'focalplane',
                                            'fiberpos.fits')
-        # with fits.open(self._fiberpos_file) as hdulist:
-        #     self.fiberpos = hdulist[1].data
-        self.fiberpos = fits.getdata(self._fiberpos_file)
+        with fits.open(self._fiberpos_file) as hdulist:
+            self.fiberpos = hdulist[1].data
 
     def _check_radec(self, ra, dec):
         """Raise ValueError if RA or dec are out of bounds.