+++ py/desimodel/test/test_trim.py
@@ -27,5 +27,8 @@ class TestTrim(unittest.TestCase):
         self.assertTrue((i1 == i2).all())
 
 
-if __name__ == '__main__':
-    unittest.main()
+def test_suite():
+    """Allows testing of only this module with the command::
+        python setup.py test -m <modulename>
+    """
+    return unittest.defaultTestLoader.loadTestsFromName(__name__)