+++ doc/release_notes.rst
@@ -19,10 +19,14 @@ desimodel Release Notes
 * Moved test of focalplane code into the actual test suite.
 * Preparing for Python 3.
 * Changed default svn version to trunk and added error handling to
-  install_desimodel_data.
+  :command:`install_desimodel_data`.
 * Update template module file to reflect DESI+Anaconda infrastructure.
-* Add code to generate random centroid offsets in desimodel.focalplane.
-* Add jupyter notebook documenting new throughput files of PR#29.
+* Add code to generate random centroid offsets in :mod:`desimodel.focalplane`.
+* Add jupyter notebook documenting new throughput files of `PR#29`_.
+* Use Astropy-recommended method of reading FITS data tables.
+* Remove reference to Travis scripts in MANIFEST.in.
+
+.. _`PR#29`: https://github.com/desihub/desimodel/pull/29
 
 0.4.5 (2016-07-15)
 ------------------